backend/services/user_service/requirements.txt
# FastAPI и веб-сервер
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Базы данных
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Аутентификация и безопасность
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Валидация данных
pydantic==2.5.0
pydantic-settings==2.1.0

# Тестирование
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Инструменты разработки
python-dotenv==1.0.0

# Логирование
structlog==23.2.0